requirements.txt
beautifulsoup4>=4.12.2
lxml>=5.2.1
python-dotenv>=1.0.1
certifi>=2024.8.30


